+++ CHANGELOG.md
@@ -17,6 +17,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2183 Merging monitoring alerts into 1 column in the patient list (step 1)
 - YLP-2184 Merging monitoring alerts into 1 column in the patient list (step 2)
 - YLP-2185 Merging monitoring alerts into 1 column in the patient list (step 3)
+- YLP-2188 Sort patients by column on patient list
 - YLP-2192 Team scope dropdown for HCP
 - YLP-2194 Navigation tabs for HCP ("Patients", "Care teams")
 - YLP-2215 Implement specific daily visualisation of UMM bolus